+++ frontend/src/components/CompletedPairRoom/ReferenceCard/ReferenceList/ReferenceList.styles.ts
@@ -44,3 +44,82 @@ export const EmptyLayout = styled.div`
   color: ${({ theme }) => theme.color.black[60]};
   font-size: ${({ theme }) => theme.fontSize.md};
 `;
+
+export const Item = styled.div`
+  display: flex;
+  flex-direction: column;
+
+  position: relative;
+
+  width: 17rem;
+  height: 20rem;
+  border: 1px solid ${({ theme }) => theme.color.black[30]};
+  border-radius: 1.5rem;
+`;
+
+export const Image = styled.img`
+  width: 100%;
+  height: 10rem;
+
+  object-fit: cover;
+  border-top-left-radius: 1.5rem;
+  border-top-right-radius: 1.5rem;
+`;
+
+export const NoneImage = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  width: 100%;
+  height: 10rem;
+
+  background-color: ${({ theme }) => theme.color.black[40]};
+  color: ${({ theme }) => theme.color.black[70]};
+  font-size: ${({ theme }) => theme.fontSize.sm};
+  line-height: 1.3;
+
+  border-top-left-radius: 1.5rem;
+  border-top-right-radius: 1.5rem;
+`;
+
+export const Box = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 0.6rem;
+  overflow: hidden;
+
+  width: 100%;
+  height: 10rem;
+  max-height: 12rem;
+  padding: 1.5rem;
+
+  cursor: pointer;
+`;
+
+export const Title = styled.p`
+  overflow: hidden;
+
+  width: 100%;
+
+  font-size: ${({ theme }) => theme.fontSize.md};
+  font-weight: ${({ theme }) => theme.fontWeight.medium};
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  word-break: break-all;
+`;
+
+export const Content = styled.p`
+  display: -webkit-box;
+  overflow: hidden;
+
+  color: ${({ theme }) => theme.color.black[60]};
+  font-size: ${({ theme }) => theme.fontSize.xs};
+  line-height: 1.5;
+  text-overflow: ellipsis;
+  white-space: normal;
+  word-break: break-all;
+
+  -webkit-box-orient: vertical;
+  -webkit-line-clamp: 3;
+`;
