+++ frontend/src/components/PairRoom/ReferenceCard/AddReferenceForm/AddReferenceForm.tsx
@@ -2,7 +2,7 @@ import { LuPlus } from 'react-icons/lu';
 
 import Button from '@/components/common/Button/Button';
 import Input from '@/components/common/Input/Input';
-import CategoryDropdown from '@/components/PairRoom/ReferenceCard/AddReferenceForm/CategoryDropdown/CategoryDropdown';
+import CategoryDropdown from '@/components/PairRoom/ReferenceCard/CategoryDropdown/CategoryDropdown';
 import { Category } from '@/components/PairRoom/ReferenceCard/ReferenceCard.type';
 
 import useInput from '@/hooks/common/useInput';
@@ -11,13 +11,13 @@ import useReference from '@/hooks/PairRoom/useReference';
 import * as S from './AddReferenceForm.styles';
 
 interface ReferenceFormProps {
-  getCategoryNameById: (categoryId: string) => string;
   categories: Category[];
   accessCode: string;
   isCategoryExist: (categoryName: string) => boolean;
+  getCategoryNameById: (categoryId: string) => string;
 }
 
-const AddReferenceForm = ({ accessCode, categories, getCategoryNameById, isCategoryExist }: ReferenceFormProps) => {
+const AddReferenceForm = ({ accessCode, categories, isCategoryExist, getCategoryNameById }: ReferenceFormProps) => {
   const { value, status, message, handleChange, resetValue } = useInput();
   const { currentCategoryId, handleCurrentCategory, handleSubmit } = useReference(accessCode, value, () =>
     resetValue(),
@@ -26,10 +26,10 @@ const AddReferenceForm = ({ accessCode, categories, getCategoryNameById, isCateg
   return (
     <S.Layout>
       <CategoryDropdown
-        currentCategoryId={currentCategoryId}
-        categories={categories}
-        handleCurrentCategory={handleCurrentCategory}
         accessCode={accessCode}
+        categories={categories}
+        currentCategoryId={currentCategoryId}
+        handleCurrentCategory={handleCurrentCategory}
         getCategoryNameById={getCategoryNameById}
         isCategoryExist={isCategoryExist}
       />
