+++ frontend/src/components/PairRoom/ReferenceCard/CategoryDropdown/CategoryDropdown.tsx
@@ -13,8 +13,8 @@ import { useAddCategory } from '@/queries/PairRoom/category/mutation';
 import { validateCategoryName } from '@/validations/validateCategory';
 
 interface CategoryDropdownProps {
-  categories: Category[];
   accessCode: string;
+  categories: Category[];
   currentCategoryId: string | null;
   handleCurrentCategory: (category: string) => void;
   getCategoryNameById: (categoryId: string) => string;
@@ -29,13 +29,23 @@ const CategoryDropdown = ({
   getCategoryNameById,
   isCategoryExist,
 }: CategoryDropdownProps) => {
-  const { value, status, message, handleChange, resetValue } = useInput();
   const { addToast } = useToastStore();
 
-  const addCategory = useAddCategory().mutateAsync;
-  const handleCategory = (option: string) => {
-    handleCurrentCategory(option);
+  const { value, status, message, handleChange, resetValue } = useInput();
+
+  const { mutateAsync } = useAddCategory();
+
+  const handleSubmit = (event: React.FormEvent) => {
+    event.preventDefault();
+
+    if (status === 'ERROR') {
+      addToast({ status, message });
+      return;
+    }
+
+    mutateAsync({ category: value, accessCode }).then(() => resetValue());
   };
+
   return (
     <Dropdown
       width="17rem"
@@ -44,26 +54,17 @@ const CategoryDropdown = ({
       placeholder="카테고리를 선택해주세요."
       valueOptions={categories}
       selectedOption={getCategoryNameById(currentCategoryId || '') || DEFAULT_CATEGORY_VALUE}
-      onSelect={(option) => handleCategory(option)}
+      onSelect={(option) => handleCurrentCategory(option)}
     >
-      <form
-        onSubmit={(event) => {
-          event.preventDefault();
-          if (status === 'ERROR') {
-            addToast({ status, message });
-            return;
-          }
-          addCategory({ category: value, accessCode }).then(() => resetValue());
-        }}
-      >
+      <form onSubmit={handleSubmit}>
         <Input
-          value={value}
-          onChange={(event) => handleChange(event, validateCategoryName(event.target.value, isCategoryExist))}
-          maxLength={15}
-          height="4rem"
-          status={status}
-          placeholder="+ 새 카테고리 추가"
           autoFocus
+          placeholder="+ 새 카테고리 추가"
+          height="4rem"
+          maxLength={15}
+          value={value}
+          status={status}
+          onChange={(event) => handleChange(event, validateCategoryName(event.target.value, isCategoryExist))}
         />
       </form>
     </Dropdown>
