+++ frontend/src/components/PairRoom/ReferenceCard/ReferenceCard.tsx
@@ -42,9 +42,9 @@ const ReferenceCard = ({ accessCode, isOpen, toggleIsOpen }: ReferenceCardProps)
             <ReferenceList references={references} accessCode={accessCode} />
             <S.Footer>
               <AddReferenceForm
-                isCategoryExist={isCategoryExist}
                 accessCode={accessCode}
                 categories={categories}
+                isCategoryExist={isCategoryExist}
                 getCategoryNameById={getCategoryNameById}
               />
             </S.Footer>
