+++ frontend/src/components/PairRoom/ReferenceCard/ReferenceList/ReferenceList.styles.ts
@@ -1,14 +1,25 @@
+import { MdClose } from 'react-icons/md';
 import styled, { css } from 'styled-components';
 
 export const Layout = styled.div<{ $columns: number }>`
   display: flex;
   flex-grow: 1;
   flex-direction: column;
-  align-items: ${({ $columns }) => ($columns > 2 ? 'center' : '')};
+  align-items: ${({ $columns }) => $columns > 2 && 'center'};
   gap: 1rem;
   overflow-y: auto;
 
   padding: 3rem;
+`;
+
+export const EmptyLayout = styled.div`
+  flex-grow: 1;
+
+  height: 0;
+  padding: 2rem;
+
+  color: ${({ theme }) => theme.color.black[60]};
+  font-size: ${({ theme }) => theme.fontSize.md};
 `;
 
 export const List = styled.ul<{ $columns: number }>`
@@ -35,12 +46,102 @@ export const List = styled.ul<{ $columns: number }>`
   }
 `;
 
-export const EmptyLayout = styled.div`
-  flex-grow: 1;
+export const Item = styled.div`
+  display: flex;
+  flex-direction: column;
 
-  height: 0;
-  padding: 2rem;
+  position: relative;
+
+  width: 17rem;
+  height: 20rem;
+  border: 1px solid ${({ theme }) => theme.color.black[30]};
+  border-radius: 1.5rem;
+`;
+
+export const Image = styled.img`
+  width: 100%;
+  height: 10rem;
+
+  object-fit: cover;
+  border-top-left-radius: 1.5rem;
+  border-top-right-radius: 1.5rem;
+`;
+
+export const NoneImage = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  width: 100%;
+  height: 10rem;
+
+  background-color: ${({ theme }) => theme.color.black[40]};
+  color: ${({ theme }) => theme.color.black[70]};
+  font-size: ${({ theme }) => theme.fontSize.sm};
+  line-height: 1.3;
+
+  border-top-left-radius: 1.5rem;
+  border-top-right-radius: 1.5rem;
+`;
+
+export const Box = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 0.6rem;
+  overflow: hidden;
+
+  width: 100%;
+  height: 10rem;
+  max-height: 12rem;
+  padding: 1.5rem;
+
+  cursor: pointer;
+`;
+
+export const Title = styled.p`
+  overflow: hidden;
+
+  width: 100%;
+
+  font-size: ${({ theme }) => theme.fontSize.md};
+  font-weight: ${({ theme }) => theme.fontWeight.medium};
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  word-break: break-all;
+`;
+
+export const Content = styled.p`
+  display: -webkit-box;
+  overflow: hidden;
 
   color: ${({ theme }) => theme.color.black[60]};
-  font-size: ${({ theme }) => theme.fontSize.md};
+  font-size: ${({ theme }) => theme.fontSize.xs};
+  line-height: 1.5;
+  text-overflow: ellipsis;
+  white-space: normal;
+  word-break: break-all;
+
+  -webkit-box-orient: vertical;
+  -webkit-line-clamp: 3;
+`;
+
+export const DeleteButton = styled(MdClose)`
+  position: absolute;
+  top: 1rem;
+  right: 1rem;
+
+  width: 2rem;
+  height: 2rem;
+  padding: 0.3rem;
+  border-radius: 100%;
+
+  background-color: ${({ theme }) => theme.color.black[90]};
+  opacity: 0.6;
+  color: ${({ theme }) => theme.color.black[20]};
+
+  cursor: pointer;
+
+  &:hover {
+    opacity: 1;
+  }
 `;
