+++ frontend/src/components/PairRoom/ReferenceCard/ReferenceList/ReferenceList.tsx
@@ -1,34 +1,46 @@
-import Reference from '@/components/PairRoom/ReferenceCard/ReferenceList/Reference/Reference';
+import { Link } from 'react-router-dom';
 
-import type { Link } from '@/apis/referenceLink';
+import type { Reference } from '@/apis/referenceLink';
 
 import { useDeleteReferenceLink } from '@/queries/PairRoom/reference/mutation';
 
 import * as S from './ReferenceList.styles';
 
 interface ReferenceListProps {
-  references?: Link[];
   accessCode: string;
+  references?: Reference[];
 }
 
-const ReferenceList = ({ references, accessCode }: ReferenceListProps) => {
-  const deleteReference = useDeleteReferenceLink().mutate;
+const ReferenceList = ({ accessCode, references }: ReferenceListProps) => {
+  const { mutate } = useDeleteReferenceLink();
 
   if (!references || references.length < 1) return <S.EmptyLayout>저장된 링크가 없습니다.</S.EmptyLayout>;
 
+  const columns = references.length;
+
   return (
-    <S.Layout $columns={references.length}>
-      <S.List $columns={references.length}>
+    <S.Layout $columns={columns}>
+      <S.List $columns={columns}>
         {references.map((reference) => {
           return (
-            <Reference
-              key={reference.id}
-              url={reference.url}
-              image={reference.image}
-              title={reference.openGraphTitle || reference.headTitle}
-              description={reference.description}
-              onDeleteReference={() => deleteReference({ id: reference.id, accessCode })}
-            />
+            <S.Item key={reference.id}>
+              <S.DeleteButton onClick={() => mutate({ id: reference.id, accessCode })} />
+              <Link to={reference.url} target="_blank">
+                {reference.image ? (
+                  <S.Image alt="link" src={reference.image} />
+                ) : (
+                  <S.NoneImage>
+                    이미지가
+                    <br />
+                    없습니다
+                  </S.NoneImage>
+                )}
+                <S.Box>
+                  <S.Title>{reference.openGraphTitle || reference.headTitle}</S.Title>
+                  <S.Content>{reference.description}</S.Content>
+                </S.Box>
+              </Link>
+            </S.Item>
           );
         })}
       </S.List>
